requirements.txt
djangorestframework==3.13.1
Django==4.1.2
gunicorn==20.1.0
psycopg2-binary==2.9.3
